dss_logic/requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.8.1
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.4
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-phone-field==1.8.1
django-widget-tweaks==1.5.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
frozenlist==1.4.1
git-filter-repo==2.47.0
gunicorn==22.0.0
idna==3.7
multidict==6.0.5
packaging==24.0
pillow==10.3.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-decouple==3.8
pytz==2024.1
requests==2.31.0
sqlparse==0.5.0
twilio==9.0.5
urllib3==2.2.1
whitenoise==6.6.0
yarl==1.9.4
